Renamed all internal attachment variables, filenames and methods to resource
