+++ packages/app-mobile/components/screens/resourceScreenUtils.ts
